+++ db/migrate/20210402001035_data_migrate_observation_matrix_column_item_single_descriptor_type.rb
@@ -1,0 +1,13 @@
+class DataMigrateObservationMatrixColumnItemSingleDescriptorType < ActiveRecord::Migration[6.0]
+  def change
+    target = ObservationMatrixColumnItem::Single::Descriptor
+
+    ObservationMatrixColumnItem.connection.execute(
+      <<~SQL
+        UPDATE #{target.table_name}
+        SET type = '#{target.to_s}'
+        WHERE type = 'ObservationMatrixColumnItem::SingleDescriptor';
+      SQL
+    )
+  end
+end