+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 - `create_backup_directory` flag to create backup directory if it does not exist for taxonworks rake tasks requiring `backup_directory`.
 
+### Fixed
+- Missing data migration for `ObservationMatrixColumnItem::SingleDescriptor` to `ObservationMatrixColumnItem::Single:Descriptor`
+
 ### Changed
 - Changed TaxonName string for superspecies names
 - Updated y18n node package to version 4.0.1 [#2160]